Refactored energy handling
- added Forge Energy support for blocks and items
- added GregtechCE energy support for blocks and items
- added energy status for laser machines
- added rate to energy status
- added display units configuration (wip)
- added explicit EnergyWrapper from ItemEnergyWrapper and TileEntityAbstractEnergy statics
- fixed crash with NULL facing input/outputs
- fixed warp electrolyzer deleting energy containers
- fixed energy storage unable to reach maximum with flux energy
- fixed laser tree farm LUA scripts for OC not installing
- improved reactor explosion logs and CPU load
Added an onUpgraded() method to simplify code structure

+++ src/main/java/cr0s/warpdrive/data/AcceleratorSetup.java
@@ -511,8 +511,8 @@ public class AcceleratorSetup extends GlobalPosition {
 	}
 	
 	// Pseudo-API for energy
-	public int energy_getEnergyStored() {
-		int energyStored = 0;
+	public long energy_getEnergyStored() {
+		long energyStored = 0;
 		for (final TileEntityCapacitor tileEntityCapacitor : setCapacitors) {
 			energyStored += tileEntityCapacitor.energy_getEnergyStored();
 		}
